+++ geo/src/lib.rs
@@ -127,6 +127,7 @@
 //! - **[`HaversineDestination`](algorithm::haversine_destination::HaversineDestination)**:
 //! - **[`HaversineIntermediate`](algorithm::haversine_intermediate::HaversineIntermediate)**:
 //! - **`Proj`**: Project geometries with the `proj` crate
+//! - **[`ChaikinSmoothing`](algorithm::chaikin_smoothing::ChaikinSmoothing)**: Smoothen `LineString`, `Polygon`, `MultiLineString` and `MultiPolygon` using Chaikins algorithm.
 //!
 //! # Features
 //!
@@ -253,6 +254,7 @@ pub mod prelude {
     pub use crate::algorithm::translate::Translate;
     pub use crate::algorithm::vincenty_distance::VincentyDistance;
     pub use crate::algorithm::vincenty_length::VincentyLength;
+    pub use crate::algorithm::chaikin_smoothing::ChaikinSmoothing;
 }
 
 /// A common numeric trait used for geo algorithms.